requirements.txt
Appium-Python-Client==4.0.0
pytest~=8.1.1
selenium~=4.19.0